hometask9/my_first_site/requirements.txt
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
polls==0.1
psycopg2==2.8.3
pytz==2019.3
sqlparse==0.3.0
